requirements.txt
attrs>=19.3.0
boto3>=1.20.25
botocore>=1.23.25
cryptography>=36.0.1
dnspython>=2.1.0
google-api-core>=2.4.0
google-api-python-client>=2.34.0
google-auth>=2.3.3
google-auth-httplib2>=0.1.0
googleapis-common-protos>=1.54.0
google-cloud>=0.34.0
google-cloud-compute>=1.6.1
google-cloud-network-management>=1.5.4
Jinja2>=3.0.3
passlib>=1.7.4
pycryptodome>=3.12.0
pytz>=2021.3
pyvmomi>=7.0.3
requests>=2.27.1
urllib3>=1.26.7
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
azure.mgmt.network>=19.3.0
azure.mgmt.compute>=24.0.0
azure.mgmt.storage>=19.0.0
azure.mgmt.subscription>=2.0.0
ply>=3.11
pytest>=6.2.5
pytest-asyncio~=0.17,>=0.17.2
pytest-rerunfailures~=10.2
